Move twitch emote-related stuff to TwitchEmotes class
